KL-177/refactor: refactor LikeButtons
- LikeButton의 동작 로직을 훅으로 분리하였습니다.
- userData를 상위 컴포넌트로부터 전달받습니다.
- LikeButton의 이름을 PreviewLikeButton으로 변경하였습니다.
- PreviewLikeButton의 이름을 ReviewLikeButton으로 변경하였습니다.
