--- src/components/Button/LikeButton.jsx
@@ -1,74 +1,0 @@
-import React, {
-  // useState,
-  useEffect,
-} from 'react'
-import PropTypes from 'prop-types'
-import { FaHeart, FaRegHeart } from 'react-icons/fa6'
-import useUserData from '../../hooks/useUserData'
-import useKyQuery from '../../hooks/useKyQuery'
-import useProductLikeFunc from '../../hooks/useProductLikeFunc'
-import IconTextButton from './IconTextButton'
-
-function LikeButton({
-  productId,
-  // likeContent = false,
-  iconSize = '1.3rem',
-}) {
-  const { data: userData } = useUserData()
-  // const [isLiked, setIsLiked] = useState(likeContent)
-  const { data: isLiked, refetch: getLike } = useKyQuery(
-    `products/${productId}/likes`,
-    [`products/likes`, productId],
-    {
-      enabled: !!userData,
-      refetchOnWindowFocus: false,
-      initialData: { data: { isLiked: false } },
-      select: (data) => data.data.isLiked,
-    }
-  )
-  const { likeProduct, unlikeProduct } = useProductLikeFunc(productId)
-
-  useEffect(() => {
-    const fetchLikeContent = async () => {
-      try {
-        await getLike()
-      } catch (error) {
-        console.error(error)
-      }
-    }
-
-    if (!userData) return
-    fetchLikeContent()
-  }, [userData])
-
-  const handleLike = async () => {
-    if (!userData) {
-      alert('로그인이 필요합니다.')
-      return
-    }
-    if (!isLiked) await likeProduct()
-    else await unlikeProduct()
-    // setIsLiked((prev) => !prev)
-  }
-
-  const iconValue = {
-    color: isLiked ? 'red' : 'darkgray',
-    size: iconSize,
-  }
-
-  return (
-    <IconTextButton
-      value={iconValue}
-      icon={isLiked ? <FaHeart /> : <FaRegHeart />}
-      handleClick={handleLike}
-    />
-  )
-}
-
-LikeButton.propTypes = {
-  // likeContent: PropTypes.bool.isRequired,
-  productId: PropTypes.number.isRequired,
-  iconSize: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-}
-
-export default LikeButton
